Refactor leave request handling and update routes
- Updated 'Tài khoản' route to 'account' and added 'leave' route.
- Removed several routes related to 'Tuyển sinh'.
- Modified `app.tsx` to include `MyNotification` and removed `Space`.
- Adjusted layout in `avatar-dropdown.tsx`.
- Introduced `LeaveRequest` component in `index.tsx` for managing leave requests.
- Added new API functions in `leave.ts` for leave request operations.
- Changed data types for `AvailableDays`, `TotalDays`, and `MaxDays` from `long`/`int` to `double` across multiple files.
- Enhanced `LeaveRequestService` with validation and improved error handling.
- Added `Comments` property to `LeaveRequestApproveArgs`.
- Updated service interfaces and classes to reflect changes in leave request handling.

+++ THPHome/Entities/Leaves/LeaveBalance.cs
@@ -8,7 +8,7 @@ public class LeaveBalance : BaseEntity
 {
     [StringLength(256)]
     public string UserName { get; set; } = default!;
-    public long AvailableDays { get; set; }
+    public double AvailableDays { get; set; }
     [ForeignKey(nameof(LeaveType))]
     public int LeaveTypeId { get; set; }
     public int Year { get; set; }
