+++ THPHome/Interfaces/IService/ILeaves/ILeaveBalanceService.cs
@@ -6,5 +6,5 @@ namespace THPHome.Interfaces.IService.ILeaves;
 public interface ILeaveBalanceService
 {
     Task<LeaveBalance?> GetBalanceByTypeAsync(int id);
-    Task<THPResult> UpdateAvailableDaysAsync(int leaveTypeId, long totalDays);
+    Task<THPResult> UpdateAvailableDaysAsync(int leaveTypeId, double totalDays);
 }
