+++ THPHome/Services/Leaves/Args/LeaveRequestUpdateArgs.cs
@@ -5,7 +5,7 @@ namespace THPHome.Services.Leaves.Args;
 public class LeaveRequestUpdateArgs : BaseEntity
 {
     public DateTime StartDate { get; set; }
-    public long TotalDays { get; set; }
+    public double TotalDays { get; set; }
     public int LeaveTypeId { get; set; }
     public string? Reason { get; set; }
 }
